some work toward issue #41
Started to implement a linked list independently of work on measure. Once this is tested thoroughly and have worked out how to reorganise directions and expressions, this will replace the current note list as it allows for organisation by time, but also a standard pattern for adding space in between notes without a lot of pushing and popping onto a python list.