added recursive toLily method
added a method which calls the toLily method of each node, the toLily method of each item, wraps the item in the node lily and then calls each child of the node. this should make the item methods a lot cleaner as they no longer have to include things like checking how many staves there are or voices, as each voice/stave can do that for itself.