some test fixes and example code for issue #45
Did some further test fixes, also discovered issue 45 in test lines. For now I'm going with ignoring it as the lilystring output won't be altered, and we can't really do option b in the case that forwards are applied (though I guess placeholders with no duration could work as a barrier case) but may go back and look at this later.