work towards issue #51
Modified how divisions are set: underlying issue was that staves aren't all known at the point of measure being created (tho could use the "staves" tag to implement this) so it now does a check at the end of measure tag. However, "r0" is still appearing in violin score of ActorPreludeSample so not entirely fixed yet.