issue fix #39
Implemented dotted notes as lilypond was falling over because of bar checks because they weren't being handled. Also had to adjust according to whether the note was in a chord or not, as with durations on chords. Fixed another bug that I had forgotten about the layer of voices in the MeasureNode meaning that chord types were not being updated at the end of every measure being put in.