work towards issue #55
Fixed a problem in how chords are handled in that the expression addition method was messing things up: it now checks to make sure the first object is a Note, if so put it to the second child etc. Also fixed the title/meta issue whereby title and such wasn't being outputted because SetupPiece was adding everything to PieceNode not Piece item - doesn't make a lot of difference but keeps things clean and means I don't have to move a bunch of code over.