work towards issue #56
Added a new method at measure item level which calculates the maximum value in lilypond notation a bar can have for use when a measure is a rest measure. This is called at measure node level when it's "rest" value is set to true, done when a measureRest note is encountered. It will then pass in the value to voice which uses it to figure out what notation to write - this skips out the note's notation entirely: at present this also skips any directions or expressions, tho thinking I should change that as even though expressions are likely to not appear, directions like text markings could still appear.
Still not perfect: if a bar does not have a meter but is a rest measure, it will fail, so we need to have a method which calls this and applies it to future methods unless they have their own meter.