issue fix #81
Problem was in fact that for some reason, the testcase I was applying this to made the parser load the title in in chunks: every time a new chunk was added it would replace the old one, and as the end of the tag had not been hit it would never have an effect on the title of the piece. Fixed this by appending text to the value where the tag is already in the chars dictionary.