work towards issue #97
tab output seems to be getting sent to lilypond, but lilypond ignores it unless we're in a TabStaff. Vetoing tab notation anyway so it doesn't make a deal of a difference, but will try and cover drum tab while I'm here so going to update the tab checker method to check both.