issue fix #100
Problem with sharps and flats appearing as naturals was actually that the alter tag was being loaded as a string, not an int, and therefore wasn't "in" the alter dictionary within the pitch item. added a cast around the alter tag as it gets loaded in and the testcase now passes, although we need to ensure children are processed in order in the PieceNode toLily method as currently their keys could come in any order.