work towards issue #120
Working with testcase 33d because this tests OctaveShifts in exclusion, but fixed a bug in which directions weren't being added in the right place.
Next step will be to figure out a way to apply this direction to all notes in between the start and stop, in order to tell them to sort out their octave.