work towards issue #122
Added code and tests to make a volta around the previous bar - this occurs where a barline has an ending and a repeat marker. The repeat marker for that bar will be changed to repeat-barline, while the one before it has the volta markings wrapped around it. In order to change this for all circumstances the addition of repeat "forward" in the previous bar should be moved to be conditional on whether we've already encountered a forward repeat - we are checking this later in the method, but for now I'll leave it where it is