issue fix #135
Made a final change that makes 45d look exactly as expected - where barline has an ending and a repeat, check the ending is the first one. If it is, set the repeat to backward-barline, otherwise ignore it.
Will need to implement issue #134 before the warning goes away, but otherwise, good to go