work on issue #312
I haven't tested this on windows yet and there's no unit tests to confirm functionality for the helper methods, so I'm not sure this is a fix, but I've put in a proposed fix for making sure the stylesheet parser works fine with both platforms by replacing any path seps from the other platform with the ones on that platform. In theory should be fne. Also used os.path.split and os.path.join wherever possible to avoid issues with slashes, rather than using icky string splits, but as it's CSS and not just a path I can't do this 100% of the time.