added test for checking stylesheet against proven working stylesheet that doesn't work
the title seems a bit weird, but due to newline and tab character issues I've commented the test out, but added a file which is a valid windows file to compare against. tbh not a great test of functionality anyway, we should try to find a validator for qss that's not just doing a comparison of two files.